je vois le bout du tunnel

+++ gateway/src/main/java/ipl/vinci/be/gateway/models/Notification.java
@@ -10,8 +10,8 @@ import lombok.*;
 
 public class Notification {
 
-    private int user_id;
-    private int trip_id;
+    private long user_id;
+    private long trip_id;
     private String date;
     private String notification_text;
